* feature/BusCommunication:
  Change: changed debug configuration to debug te MFM sensor and MFM basis simultaneously
  Docs: added doxygen documentatie to the utils functions
  Change: moved functions out of the main to a seperate library
  Change: Reduced the modbus timeout
  Fix: Fixed slot id calculation
  Add: added conditional check that if sensor is not found during the init , it will not sample this sensor to save time
  Change: change data register form int32_t to a float. Change mainloop to facilitate this
  Add: added check on timeout of the uart function and return zero when there is no message received
  Add: added function to get both pressure and temperature in one request. Only a float can be read out in this method. Need to change the sensor data register type to contain a float.
  Change: turn on debug led on start
  Change: Changed slot gpio to default input

+++ App/i2cSlave/Src/SensorRegister.c
@@ -139,7 +139,7 @@ uint8_t readMeasSamples(void)
  * @param data is the raw sensor data
  * @param sensor is the sensor the data is taken from
  */
-void storeMeasurement(int32_t pressure, int32_t temperature, uint8_t sensor)
+void storeMeasurement(float pressure, float temperature, uint8_t sensor)
 {
   registerMeasurementData[sensor].pressure = pressure;
   registerMeasurementData[sensor].temperature = temperature;
